+++ cicdStageExecutor/StageExecutor_test.go
@@ -1,7 +1,6 @@
-package main
+package cicdStageExecutor
 
 import (
-	"github.com/devtron-labs/ci-runner/cicdStageExecutor"
 	"github.com/devtron-labs/ci-runner/helper"
 	"reflect"
 	"testing"
@@ -84,7 +83,7 @@ func Test_deduceVariables(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := cicdStageExecutor.deduceVariables(tt.args.desiredVars, tt.args.globalVars, tt.args.preeCiStageVariable, tt.args.postCiStageVariables, nil)
+			got, err := deduceVariables(tt.args.desiredVars, tt.args.globalVars, tt.args.preeCiStageVariable, tt.args.postCiStageVariables, nil)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("deduceVariables() error = %v, wantErr %v", err, tt.wantErr)
 				return
@@ -98,7 +97,7 @@ func Test_deduceVariables(t *testing.T) {
 
 func TestRunCiSteps(t *testing.T) {
 	type args struct {
-		stageType                  cicdStageExecutor.StepType
+		stageType                  StepType
 		req                        *helper.CommonWorkflowRequest
 		globalEnvironmentVariables map[string]string
 		preeCiStageVariable        map[int]map[string]*helper.VariableObject
@@ -114,7 +113,7 @@ func TestRunCiSteps(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			gotPreeCiStageVariableOut, gotPostCiStageVariable, err := cicdStageExecutor.RunCiCdSteps(tt.args.stageType, tt.args.req.PreCiSteps, nil, tt.args.globalEnvironmentVariables, tt.args.preeCiStageVariable)
+			gotPreeCiStageVariableOut, gotPostCiStageVariable, err := RunCiCdSteps(tt.args.stageType, tt.args.req.PreCiSteps, nil, tt.args.globalEnvironmentVariables, tt.args.preeCiStageVariable)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("RunCiCdSteps() error = %v, wantErr %v", err, tt.wantErr)
 				return
