+++ cicdStageExecutor/TaskExecutor.go
@@ -15,6 +15,4 @@
  *
  */
 
-package main
-
-
+package cicdStageExecutor
