+++ cicdStageExecutor/scriptExecutor_test.go
@@ -1,4 +1,4 @@
-package main
+package cicdStageExecutor
 
 import (
 	"fmt"
@@ -62,7 +62,7 @@ func TestRunScripts(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := ciCdStageExecutor.RunScripts(tt.args.workDirectory, tt.args.scriptFileName, tt.args.script, tt.args.envVars, tt.args.outputVars)
+			got, err := RunScripts(tt.args.workDirectory, tt.args.scriptFileName, tt.args.script, tt.args.envVars, tt.args.outputVars)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("RunScripts() error = %v, wantErr %v", err, tt.wantErr)
 				return
@@ -76,7 +76,7 @@ func TestRunScripts(t *testing.T) {
 
 func Test_buildDockerRunCommand(t *testing.T) {
 	type args struct {
-		executionConf *ciCdStageExecutor.executionConf
+		executionConf *executionConf
 	}
 	tests := []struct {
 		name    string
@@ -85,7 +85,7 @@ func Test_buildDockerRunCommand(t *testing.T) {
 		wantErr bool
 	}{
 		{name: "all_single",
-			args: args{executionConf: &ciCdStageExecutor.executionConf{
+			args: args{executionConf: &executionConf{
 				DockerImage:         "alpine:latest",
 				EnvInputFileName:    "/tmp/ci-test/abc.env",
 				EntryScriptFileName: "/tmp/code-location/_entry.sh",
@@ -99,7 +99,7 @@ func Test_buildDockerRunCommand(t *testing.T) {
 			want:    "docker run --network host \\\n--env-file /tmp/ci-test/abc.env \\\n-v /tmp/code-location/_entry.sh:/devtron_script/_entry.sh \\\n-v /tmp/ci-test/_env.out:/devtron_script/_out.env \\\n-v /tmp/code-location:/tmp/code-mount-location \\\n-v /src:/des \\\n-v /tmp/custom-script-location:/tmp/script-mount-location \\\n-p 80:8080 \\alpine:latest \\\n/bin/sh /devtron_script/_entry.sh\n",
 		},
 		{name: "all_multi",
-			args: args{executionConf: &ciCdStageExecutor.executionConf{
+			args: args{executionConf: &executionConf{
 				DockerImage:         "alpine:latest",
 				EnvInputFileName:    "/tmp/ci-test/abc.env",
 				EntryScriptFileName: "/tmp/code-location/_entry.sh",
@@ -115,7 +115,7 @@ func Test_buildDockerRunCommand(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := ciCdStageExecutor.buildDockerRunCommand(tt.args.executionConf)
+			got, err := buildDockerRunCommand(tt.args.executionConf)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("buildDockerRunCommand() error = %v, wantErr %v", err, tt.wantErr)
 				return
@@ -158,7 +158,7 @@ func Test_buildDockerEntryScript(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := ciCdStageExecutor.buildDockerEntryScript(tt.args.command, tt.args.args, tt.args.outputVars)
+			got, err := buildDockerEntryScript(tt.args.command, tt.args.args, tt.args.outputVars)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("buildDockerEntryScript() error = %v, wantErr %v", err, tt.wantErr)
 				return
@@ -174,7 +174,7 @@ func Test_buildDockerEntryScript(t *testing.T) {
 func TestRunScriptsInDocker(t *testing.T) {
 	t.SkipNow()
 	type args struct {
-		executionConf *ciCdStageExecutor.executionConf
+		executionConf *executionConf
 	}
 	tests := []struct {
 		name    string
@@ -184,7 +184,7 @@ func TestRunScriptsInDocker(t *testing.T) {
 	}{
 		{name: "hello",
 			args: args{
-				executionConf: &ciCdStageExecutor.executionConf{
+				executionConf: &executionConf{
 					Script:            "ls",
 					EnvInputVars:      map[string]string{"KIND": "TEST"},
 					ExposedPorts:      map[int]int{80: 8080, 90: 9090},
@@ -203,7 +203,7 @@ func TestRunScriptsInDocker(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := ciCdStageExecutor.RunScriptsInDocker(tt.args.executionConf)
+			got, err := RunScriptsInDocker(tt.args.executionConf)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("RunScriptsInDocker() error = %v, wantErr %v", err, tt.wantErr)
 				return
@@ -257,7 +257,7 @@ func Test_writeToEnvFile(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			err := ciCdStageExecutor.writeToEnvFile(tt.args.envMap, tt.args.filename)
+			err := writeToEnvFile(tt.args.envMap, tt.args.filename)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("writeToEnvFile() error = %v, wantErr %v", err, tt.wantErr)
 				return
